+++ src/parser/mod.rs
@@ -34,7 +34,7 @@ use self::util::{
 use self::value::{
     dictionary, function_call, single_value, value_expression,
 };
-use crate::sass::{FormalArgs, Item, Name, Selectors, Value};
+use crate::sass::{Callable, FormalArgs, Item, Name, Selectors, Value};
 use crate::value::ListSeparator;
 #[cfg(test)]
 use crate::value::{Numeric, Rgba, Unit};
@@ -273,11 +273,25 @@ fn mixin_call2(input: Span) -> PResult<Item> {
     let (rest, n2) = opt(preceded(tag("."), name))(rest)?;
     let name = n2.map(|n2| format!("{}.{}", n1, n2)).unwrap_or(n1);
     let (rest, _) = opt_spacelike(rest)?;
-    let (rest, args) = terminated(opt(call_args), opt_spacelike)(rest)?;
-    let (end, body) = terminated(
-        opt(body_block),
-        terminated(opt_spacelike, opt(tag(";"))),
-    )(rest)?;
+    let (rest0, args) = terminated(opt(call_args), opt_spacelike)(rest)?;
+    let (rest, t) = alt((tag("using"), tag("{"), tag("")))(rest0)?;
+    let (end, body) = match *t.fragment() {
+        b"using" => {
+            let (end, args) = preceded(opt_spacelike, formal_args)(rest)?;
+            let (rest, body) = preceded(opt_spacelike, body_block)(end)?;
+            let decl = SourcePos::from_to(rest0, end);
+            (rest, Some(Callable::new(args, body, decl)))
+        }
+        b"{" => {
+            let (rest, body) = body_block(rest0)?;
+            let decl = SourcePos::from_to(rest0, rest);
+            (rest, Some(Callable::no_args(body, decl)))
+        }
+        _ => {
+            let (rest, _) = opt(tag(";"))(rest)?;
+            (rest, None)
+        }
+    };
     let pos = SourcePos::from_to(input, rest).opt_back("@include ");
     Ok((
         end,
@@ -503,16 +517,22 @@ fn mixin_declaration2(input: Span) -> PResult<Item> {
     let (rest, args) = opt(formal_args)(rest)?;
     let (end, body) = preceded(opt_spacelike, body_block)(rest)?;
     let args = args.unwrap_or_else(FormalArgs::none);
-    let pos = SourcePos::from_to(input, rest);
-    Ok((end, Item::MixinDeclaration(name, args, body, pos)))
+    let decl = SourcePos::from_to(input, rest);
+    Ok((
+        end,
+        Item::MixinDeclaration(name, Callable { args, body, decl }),
+    ))
 }
 
 fn function_declaration2(input: Span) -> PResult<Item> {
     let (end, name) = terminated(name, opt_spacelike)(input)?;
     let (end, args) = formal_args(end)?;
     let (rest, body) = preceded(opt_spacelike, body_block)(end)?;
-    let pos = SourcePos::from_to(input, end);
-    Ok((rest, Item::FunctionDeclaration(name, args, pos, body)))
+    let decl = SourcePos::from_to(input, end);
+    Ok((
+        rest,
+        Item::FunctionDeclaration(name, Callable { args, body, decl }),
+    ))
 }
 
 fn return_stmt2<'a>(input0: Span<'_>, input: Span<'a>) -> PResult<'a, Item> {
@@ -526,9 +546,10 @@ fn return_stmt2<'a>(input0: Span<'_>, input: Span<'a>) -> PResult<'a, Item> {
 /// The "rest" of an `@content` statement is just an optional terminator
 fn content_stmt2(input: Span) -> PResult<Item> {
     let (rest, _) = opt_spacelike(input)?;
+    let (rest, args) = opt(call_args)(rest)?;
     let (rest, _) = opt(tag(";"))(rest)?;
     let pos = SourcePos::from_to(input, rest);
-    Ok((rest, Item::Content(pos)))
+    Ok((rest, Item::Content(args.unwrap_or_default(), pos)))
 }
 
 fn custom_property(input: Span) -> PResult<Item> {