+++ src/sass/functions/meta.rs
@@ -3,7 +3,7 @@ use super::{
     CheckedArg, Error, FunctionMap,
 };
 use crate::css::{CallArgs, CssString, Value};
-use crate::sass::{Function, MixinDecl, Name};
+use crate::sass::{Call, Function, MixinDecl};
 use crate::value::Quotes;
 use crate::{Format, Scope, ScopeRef};
 
@@ -60,14 +60,16 @@ pub fn create_module() -> Scope {
         };
         let args = CallArgs::from_value(s.get(&name!(args))?)?;
         if let Some(function) = function {
-            function.call(call_scope(s), args)
+            function.call(Call {
+                args,
+                scope: call_scope(s),
+            })
         } else {
             Ok(Value::Call(name, args))
         }
     });
     def!(f, content_exists(), |s| {
-        let content = call_scope(s).get_mixin(&Name::from_static("%%BODY%%"));
-        if let Some(content) = content {
+        if let Some(content) = call_scope(s).get_content() {
             Ok((!content.is_no_body()).into())
         } else {
             Err(Error::error(