+++ src/sass/mixin.rs
@@ -1,5 +1,5 @@
 use super::functions::get_string;
-use super::{CallArgs, FormalArgs, Item, Name, Value};
+use super::{CallArgs, Callable, Closure, FormalArgs, Name, Value};
 use crate::css::{CssString, ValueToMapError};
 use crate::file_context::FileContext;
 use crate::ordermap::OrderMap;
@@ -11,43 +11,20 @@ use std::convert::TryInto;
 #[derive(Clone)]
 pub enum MixinDecl {
     /// an actual mixin
-    Sass(MixinDeclImpl),
+    Sass(Closure),
     /// The body of a mixin call that does not have a body
     NoBody,
     /// The special `load-css` mixin.
     LoadCss,
 }
 
-#[derive(Clone)]
-pub struct MixinDeclImpl {
-    args: FormalArgs,
-    scope: ScopeRef,
-    body: Vec<Item>,
-    pos: SourcePos,
-}
-
-impl From<MixinDeclImpl> for MixinDecl {
-    fn from(decl: MixinDeclImpl) -> Self {
+impl From<Closure> for MixinDecl {
+    fn from(decl: Closure) -> Self {
         MixinDecl::Sass(decl)
     }
 }
 
 impl MixinDecl {
-    /// Create a new Mixin.
-    pub fn new(
-        args: &FormalArgs,
-        scope: &ScopeRef,
-        body: &[Item],
-        pos: &SourcePos,
-    ) -> Self {
-        MixinDeclImpl {
-            args: args.clone(),
-            scope: scope.clone(),
-            body: body.to_vec(),
-            pos: pos.clone(),
-        }
-        .into()
-    }
     pub(crate) fn get(
         self,
         name: &str,
@@ -61,21 +38,22 @@ impl MixinDecl {
                 let sel = scope.get_selectors().clone();
                 Ok(Mixin {
                     scope: decl
+                        .body
                         .args
-                        .eval(
+                        .evalcall(
                             ScopeRef::sub_selectors(decl.scope.clone(), sel),
                             call_args.evaluate(scope)?,
                         )
                         .map_err(|e| {
                             e.decl_called(
                                 call_pos.in_call(name),
-                                decl.pos.clone(),
+                                decl.body.decl.clone(),
                             )
                         })?,
-                    body: Parsed::Scss(decl.body),
+                    body: Parsed::Scss(decl.body.body),
                 })
             }
-            MixinDecl::NoBody => unreachable!(),
+            MixinDecl::NoBody => Ok(Mixin::empty(scope)),
             MixinDecl::LoadCss => {
                 let fargs = FormalArgs::new(vec![
                     (name!(url), None),
@@ -88,7 +66,10 @@ impl MixinDecl {
                 );
                 let call_pos2 = call_pos.clone();
                 let argscope = fargs
-                    .eval(scope.clone(), call_args.evaluate(scope.clone())?)
+                    .evalcall(
+                        scope.clone(),
+                        call_args.evaluate(scope.clone())?,
+                    )
                     .map_err(|e| e.decl_called(call_pos2, pos))?;
                 let call_pos2 = call_pos.clone();
                 let url = get_string(&argscope, name!(url)).map_err(|e| {
@@ -103,10 +84,7 @@ impl MixinDecl {
                     .any(|name| *name == url.value())
                 {
                     if with.unwrap_or_default().is_empty() {
-                        return Ok(Mixin {
-                            scope,
-                            body: Parsed::Css(vec![]),
-                        });
+                        return Ok(Mixin::empty(scope));
                     } else {
                         return Err(Error::BadCall(
                             format!(
@@ -157,25 +135,21 @@ pub struct Mixin {
 }
 
 impl Mixin {
+    fn empty(scope: ScopeRef) -> Self {
+        Mixin {
+            scope,
+            body: Parsed::Css(vec![]),
+        }
+    }
     pub(crate) fn define_content(
         &self,
         scope: &ScopeRef,
-        body: &Option<Vec<Item>>,
-        pos: &SourcePos,
+        body: &Option<Callable>,
     ) {
-        self.scope.define_mixin(
-            Name::from_static("%%BODY%%"),
-            match body {
-                Some(body) => MixinDeclImpl {
-                    args: FormalArgs::none(),
-                    scope: scope.clone(),
-                    body: body.to_vec(),
-                    pos: pos.clone(),
-                }
-                .into(),
-                None => MixinDecl::NoBody,
-            },
-        )
+        self.scope.define_content(match body {
+            Some(body) => body.closure(scope).into(),
+            None => MixinDecl::NoBody,
+        })
     }
 }